+++ packages/learn-card-init/CHANGELOG.md
@@ -1,5 +1,33 @@
 # learn-card-core
 
+## 1.2.33
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/learn-cloud-plugin@2.0.6
+
+## 1.2.32
+
+### Patch Changes
+
+-   Updated dependencies [[`c148309`](https://github.com/learningeconomy/LearnCard/commit/c148309c0dbb8d512c40a1b1f16c0e843f3e1aff)]:
+    -   @learncard/network-plugin@1.8.3
+
+## 1.2.31
+
+### Patch Changes
+
+-   Updated dependencies [[`b6dd8b6`](https://github.com/learningeconomy/LearnCard/commit/b6dd8b65e295e0fe5122658cd351e1561d541523)]:
+    -   @learncard/network-plugin@1.8.2
+
+## 1.2.30
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/network-plugin@1.8.1
+
 ## 1.2.29
 
 ### Patch Changes